Add fold to STklos core
reduce/fold are defined in logical.stk, and SRFI 224, even
existing in SRFI-1. It would be nice to have them in the core.

This PR adds `fold` and makes SRFI-1 use and reexport it.

Of course, since it is implemented in C, there is a big speedup.

```
(let ((L (iota 10000)))
  (time
    (repeat 10000
      (fold + 0 L)) -1))
```

Old code: 7116.086 ms, 1360000 bytes in 30000 allocation calls
New code: 2374.2 ms, NO allocations

Passes all tests (including the SRFI-1 tests for fold).