Optimzie dec!
inc! dec! pop! and push! are build by the same procedure, %make-inc-dec-push-body.
But dec! and push! operate in different ways:

```
(dec!  a 2)  =>  (set! a (-    a 2)) ;; "a" and "2" in the order they appeared...
(push! a 2)  =>  (set! a (cons 2 a)) ;; "a" and "2" in reverse order!
```

So the previous solution was to use a lambda to exchange the arguments in dec!.
However, this makes it slow:

```
(let ((a 1)) (time (repeat 100000000 (inc! a))) a)
;    Elapsed time: 1490.876 ms
```

```
(let ((a 1)) (time (repeat 100000000 (dec! a))) a)
;    Elapsed time: 3871.428 ms
```

So... We can change the procedure %make-inc-dec-push-body so it
takes na extra optional argument that tells wether the arguments
should be in reversed order.

After this patch, the timing for dec! is the same for inc!, since
their definitions become quite similar:

```
(define-macro (inc! place :optional (value 1))
  (%make-inc-dec-push-body place '+ value))

(define-macro (dec! place :optional (value 1))
  (%make-inc-dec-push-body place '- value #f))
```