Used a pool of unused cells to decrease the number of GC
To avoid to "abandon" unused objects when the function returns, we keep
the (already allocated, but unused, cells, in the `pool` variable. When the
function is entered again, it will be used, in place of a call to the function
`STk_must_malloc_many`. Hence, this reduces tremendously the work of