Optimize with-hander
Internally `with-handler` use a sigsetjmp/siglongjmp .  In fact it is
not really a sigsetjmp but an ad-hoc version which use setjmp and
logjmp with our version to grab the signal masks and restore
them (sigsetjmp is Posix 2008, and code was written before they were
standard). Our version for that were called get_signal_mask /
set_signal_mask.

This functions are very expensive: a simple program consisting
of a loop (10_000_000 iterations) around get_signal_mask takes 10s on
my machine.  The size of a sigset_t is 128 bytes (yes 128, that is
crazy, it was 4 some years before), It explains why this is slow.  If
we suppress the saving/restoring of signal mask, the tests still run
without errors. That seems to solve the problem, but I'm sure that I
was constrained to save the signal masks, but can remember
why (something with threads, but not sure).

Code uses macros around jmp_buf/setjmp/longjmp. They could be replaced
by their sigXXX counterpart.

This should fix the Issue #753