NEWCELL uses now a function with a pool of pre-allocatted cells
NEWCELL calls a new allocating function `STk_must_malloc_cell`
(this is the only modification to NEWCELL, and this is the only place
it will be called). We take the asked size and divide it by the size
of a pointer. It will give us the index in the pool (an array of lists
of (eventually) pre-allocated objects (the object that remains from a
previous `STk_must_malloc_many`)). Only small objects use the pool.