Ensure we are running libreadline to run the startup hook
The primitive readline-set-option is non-effective with libedit but it can be
called. However, it seems to not work when called from the
readline-startup-hook (primitive is not defined in this case). But it works ,
if run in the debugger :-<
Really, I do not understand where is the problem here

Solution: run the hook only if we are sure that the library used is readline.