Optimize length+, dotted-list? and iota in SRFI-1
1. Optimize length+, dotted-list?

Since we already have the C function `list_type_and_length()` in
STklos core, it makes no sense to keep these in Scheme.

2. Optimize iota

Because it's easy. And we can allocate the full list at once now.
The code in SRFI-1 wqill do N allocations...

BUT

We had to declare list_type_and_length non-static and include it
in stklos.h...