Rewrite drop, take & similar ones in C
Benchmarks for the following code:

```
(let ((x (append (iota 5000)
                 (iota 5000))))
  (time
    (repeat 5000 (PROC x 5000))))
```

Iota was called twice because the old code uses the stack and
STklos crashes if we allocate all at once (the new version of
iota fixes that).

In the following table, (t) means time, (a) means allocations.
For allocations "a/b" means "a bytes in b allocation calls".
The code is the same, only the procedure ("PROC") changes.

|proc            | old code           | new code       |
|----------------|--------------------|----------------|
|take (t)        | 1961.631           | 541.774        |
|take (a)        | 600600000/25015000 | 600000/5000    |
|take! (t)       | 1104.309           | 68.052         |
|take! (a)       | 600000/15000       | 0/0            |
|drop  (t)       | 1105.489           | 61.307         |
|drop  (a)       | 600000/15000       | 0/0            |
|take-right (t)  | 2135.459           | 194.576        |
|take-right (a)  | 1200000/30000      | 0/0            |
|drop-right  (t) | 2949.675           | 698.313        |
|drop-right  (a) | 601200000/25030000 | 600000000/5000 |
|drop-right! (t) | 1164.039           | 94.774         |
|drop-right! (a) | 600152/15004       | 0/0            |

Passes all tests... :)