Code cleaning
We have now all special calls implemented with the new internal-call type
`(tc_icall`). This should resolve Issue #801.

An internal call is treated specially by the compiler and generally does not
produce a function call, The implementation of the function call is done by the
function `compile-%%foo` in the file `compiler.stk`.
For instance,

   (if cond a b)

is transformed in
   (#%if cond a b)

by the `if` macro. When the compiler sees thif fom it calls the compiler
function `compiler-%%foo`.

Internal function call `#%if` is written as `#%if` and read back as `#%if`.

NOTE: I have tried an implementation with non interned symbols, but the
problem is that non interned symbols are read back as interned ones , when
written in the constants of an '.ostk' file.

Modified-by: Erick Gallesio <eg@stklos.net>