WIP: New type of object for environment
An environment objects grabs the static and dynamic environment in a
point of a program (code on on static environmenet will come later).
The static environent is more or less the value of env parameter in
the compiler. It permits the compilation of an expression in a given
context. The dynamic environment permits its execution with eval.