Fix fxsqrt and optimize exact-integer-sqrt
stklos> (fxsqrt 1152921504606846753)
1073741824
-223

Oops...

This was happening because we were not being cautious with the result from libc sqrt. It's fixed.

While we're at it, why not make exact-integer-sqrt use fxsqrt when the argument is a fixnum?
And use GMP for bignums?

```scheme
(let
  ((k (- (expt 2 60) 223)))
    (time
      (repeat 5_000_000
        (exact-integer-sqrt k))))
Old code: 2148.563 ms, Allocations: 720000000 bytes
New code: 161.115 ms, Allocations: 0
```