Use BIGNUM_FITS_DOUBLE instead of BIGNUM_FITS_INTEGER
We were doing BIGNUM_FITS_INTEGER for testing wether a GMP bignum can
be stored in a double. This patch makes log and exact->inexact
faster when applied to bignums that are in the range that was
excluded (the largest fixnum and the largest double).

In the following examples, LARGE is
```
(* 9007199254740991 (expt 2 971))
```
So:
```
(time (repeat 10_000_000 (inexact LARGE)))
Old code: 1203.126 ms
New code:  785.996 ms

(time (repeat 100_000_000 (log LARGE)
Old code: 1204.692 ms
New code:  698.890 ms
```