(expt x bignum) is zero if -1.0 < x < 1.0 and x is inexact
STklos previously complained ("exponent too big") when the
exponent was a bignum. But an inexact number within the range
(-1.0, 1.0) would result in something so small that we can
consider it to be zero.

Similar behavior can be seen in Gauche, Gambit, Chez, Guile,
Chibi (they all return positive zero when x is inexact and
the exponent is a bignum).

Also handle some other similar cases, like (expt -1 bignum) etc.

Some tests included.